Moved function to common.py
