+++ datalight/common.py
@@ -1,9 +1,9 @@
 """This module is common core functions for datalight."""
 
-import os
 import sys
 import zipfile
 import pathlib
+import configparser
 
 
 class DatalightException(Exception):
@@ -20,11 +20,12 @@ def get_files_path(directory_name):
     :raises FileNotFoundError: if the directory does not exist.
     """
 
+    directory_name = pathlib.Path(directory_name)
     # If directory_name is a file, return a list with file_name
-    if os.path.isfile(directory_name):
-        file_paths = [pathlib.Path(directory_name)]
+    if directory_name.is_file():
+        file_paths = [directory_name]
     else:
-        file_paths = pathlib.Path(directory_name).glob('**/*')
+        file_paths = directory_name.glob('**/*')
 
     absolute_paths = []
 
@@ -46,7 +47,7 @@ def zip_data(files, zip_name):
     :param zip_name: (str) Name of the zip file to create.
     """
 
-    zip_name = os.path.abspath(zip_name)
+    zip_name = pathlib.Path(zip_name).resolve()
 
     try:
         with zipfile.ZipFile(zip_name, 'w') as output_zip:
@@ -55,3 +56,29 @@ def zip_data(files, zip_name):
     except FileExistsError:
         print("Error: Zip file \"{}\" already exists. Cannot overwrite.".format(zip_name))
         sys.exit()
+
+
+def get_authentication_token(sandbox):
+    current_directory = pathlib.Path(__file__).parent
+
+    token_file = current_directory.parent / pathlib.Path("datalight.config")
+    zeno_config = configparser.ConfigParser()
+    zeno_config.read(token_file)
+    try:
+        if sandbox:
+            token = zeno_config['sandbox.zenodo.org']['token']
+        else:
+            token = zeno_config['zenodo.org']['token']
+    except KeyError:
+        token = input('Provide Zenodo token: ')
+
+        # Save the token to the ~/.zenodo
+        config = configparser.ConfigParser()
+        if sandbox:
+            config['sandbox.zenodo.org'] = {'lightform': token}
+        else:
+            config['zenodo.org'] = {'lightform': token}
+
+        with open(token_file, 'a', encoding="utf-8") as configfile:
+            config.write(configfile)
+    return token
