+++ datalight/datalight.py
@@ -2,11 +2,10 @@
 
 import os
 import sys
-import configparser
 
-from conf import logger
-from common import DatalightException, zip_data, get_files_path
-from zenodo import Zenodo as DataRepo
+from datalight.conf import logger
+from datalight.common import get_files_path, DatalightException, zip_data, get_authentication_token
+from datalight.zenodo import Zenodo as DataRepo
 
 
 def main(directory_name, metadata, zip_name="data.zip", publish=False, sandbox=True):
@@ -35,32 +34,6 @@ def main(directory_name, metadata, zip_name="data.zip", publish=False, sandbox=T
     data_repo.upload_metadata()
     if publish:
         data_repo.publish()
-
-
-def get_authentication_token(sandbox):
-    current_directory = os.path.dirname(os.path.realpath(__file__))
-
-    token_file = os.path.join(current_directory, "..", "datalight.config")
-    zeno_config = configparser.ConfigParser()
-    zeno_config.read(token_file)
-    try:
-        if sandbox:
-            token = zeno_config['sandbox.zenodo.org']['token']
-        else:
-            token = zeno_config['zenodo.org']['token']
-    except KeyError:
-        token = input('Provide Zenodo token: ')
-
-        # Save the token to the ~/.zenodo
-        config = configparser.ConfigParser()
-        if sandbox:
-            config['sandbox.zenodo.org'] = {'lightform': token}
-        else:
-            config['zenodo.org'] = {'lightform': token}
-
-        with open(token_file, 'a', encoding="utf-8") as configfile:
-            config.write(configfile)
-    return token
 
 
 if __name__ == '__main__':
